Extract KeycloakUserProvisioningService from UserService
UserService mixed local User CRUD with the raw Keycloak identity calls.
Move the three Keycloak primitives (create user + set password + assign the
default role, delete user, existence check) into a focused
KeycloakUserProvisioningService, and inject it into UserService.

The registration saga stays in UserService.registerUser: it still creates
the Keycloak user, persists the local row, and on a database failure calls
the provisioning service to roll the Keycloak user back. Behaviour is
unchanged; UserService no longer holds the Keycloak client or realm.
